+++ packages/headless/src/test/mock-insight-state.ts
@@ -4,6 +4,7 @@ import {getFacetOptionsInitialState} from '../features/facet-options/facet-optio
 import {getCategoryFacetSetInitialState} from '../features/facets/category-facet-set/category-facet-set-state';
 import {getFacetSearchSetInitialState} from '../features/facets/facet-search-set/specific/specific-facet-search-set-state';
 import {getFacetSetInitialState} from '../features/facets/facet-set/facet-set-state';
+import {getNumericFacetSetInitialState} from '../features/facets/range-facets/numeric-facet-set/numeric-facet-set-state';
 import {getDateFacetSetInitialState} from '../features/facets/range-facets/date-facet-set/date-facet-set-state';
 import {getInsightConfigurationInitialState} from '../features/insight-configuration/insight-configuration-state';
 import {getInsightInterfaceInitialState} from '../features/insight-interface/insight-interface-state';
@@ -14,6 +15,7 @@ import {getQueryInitialState} from '../features/query/query-state';
 import {getResultPreviewInitialState} from '../features/result-preview/result-preview-state';
 import {getSearchHubInitialState} from '../features/search-hub/search-hub-state';
 import {getSearchInitialState} from '../features/search/search-state';
+import {getSortCriteriaInitialState} from '../features/sort-criteria/sort-criteria-state';
 import {InsightAppState} from '../state/insight-app-state';
 
 export function buildMockInsightState(
@@ -35,8 +37,10 @@ export function buildMockInsightState(
     facetOptions: getFacetOptionsInitialState(),
     querySet: getQuerySetInitialState(),
     querySuggest: getQuerySuggestSetInitialState(),
+    numericFacetSet: getNumericFacetSetInitialState(),
     dateFacetSet: getDateFacetSetInitialState(),
     categoryFacetSet: getCategoryFacetSetInitialState(),
+    sortCriteria: getSortCriteriaInitialState(),
     ...config,
   };
 }